requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
astroid==2.15.4
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==40.0.2
curlify==2.2.1
dill==0.3.6
dj-database-url==2.0.0
Django==4.2.1
django-celery-results==2.5.1
django-cors-headers==3.14.0
django-environ==0.10.0
django-phonenumber-field==7.1.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
facebook-business==16.0.2
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
install==1.3.5
isort==5.12.0
kombu==5.2.4
lazy-object-proxy==1.9.0
mccabe==0.7.0
monday==1.3.3
multidict==6.0.4
openai==0.27.6
phonenumberslite==8.13.11
pipreqs==0.4.13
platformdirs==3.5.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pycountry==22.3.5
pycparser==2.21
PyJWT==2.7.0
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
stripe==5.4.0
tomlkit==0.11.8
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.4.0
wrapt==1.15.0
yarg==0.1.9
yarl==1.9.2
